Chore: 벡터 뺄셈 연산 분리
